Lots of progress
- Carthage frameworks
- RAC combinators
- Models
- Rough view controller stubs
- Added more unit tests
- Added a Sinatra server for hosting test data -- but needs some Travis-CI love to make it work automatically.

+++ TexLegeEngineTests/BasicBuildTests.m
@@ -25,7 +25,7 @@
 
 - (void)testIndexOf {
     NSArray *array = @[@"Greg",@"Jim",@"Alex"];
-    NSInteger index = ASTIndexOf(array, @"Alex");
+    UInt16 index = (UInt16)ASTIndexOf(array, @"Alex");
     XCTAssertEqual(index, 2);
 }
